+++ CHANGELOG.md
@@ -16,6 +16,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `dialog` conversation IO
 - `patterns` argument for adding and matching pattern layers on `simple` banner items
 - `equip` action to put items into players' equipment slots
+- `slots` to `armor` condition to only check specific slots
 ### Changed
 - `worldguard` integration version requirement to 7.0.0
 - `worldedit` integration version requirement to 7.3.0